Refactoring code to follow a more RESTful convention and fewer required inputs

+++ controllers/video-controllers.js
@@ -9,9 +9,11 @@ const uploadVideo = async (req, res) => {
 
 const stream = async (req, res) => {
     const { range } = req.headers;
-    const { fileName } = req.params;
+    const { videoID } = req.params;
     if (!range) return;
-    const videoPath = `videos/${fileName}`;
+    const video = await Video.findById(videoID);
+    if (!video) return;
+    const videoPath = `videos/${video.videoPath}`;
     const videoSize = fs.statSync(videoPath).size;
     const start = Number(range.replace(/\D/g, ""));
     const end = Math.min(start + 10**6, videoSize - 1);
@@ -22,18 +24,21 @@ const stream = async (req, res) => {
 };
 
 const updateVideo = async (req, res) => {
-    const { videoID, name } = req.body;
-    const video = await Video.findOneAndUpdate({ _id: videoID }, { name }, { runValidators: true, new: true });
+    const { videoID } = req.params;
+    const { name } = req.body
+    const video = await Video.findByIdAndUpdate(videoID, { name }, { runValidators: true, new: true });
     res.status(200).json({ status: "success", video });
 };
 
 const deleteVideo = async (req, res) => {
-    const { videoID, path } = req.params;
-    const videoPath = `videos/${path}`;
+    const { videoID } = req.params;
+    const video = await Video.findById(videoID);
+    if (!video) return;
+    const videoPath = `videos/${video.videoPath}`;
     if (!fs.existsSync(videoPath)) return;
     fs.unlink(videoPath, error => {
         if (error) return;
-        await Video.findOneAndDelete({ _id: videoID });
+        await Video.findByIdAndDelete({ _id: videoID });
         return res.status(204).json({ status: "success", video: null });
     });
 };
